Finalize sdk e2e testing environment

+++ test/e2e/runSdkTests.mjs
@@ -1,22 +1,23 @@
-import { execFile } from 'node:child_process';
+import { spawn } from 'node:child_process';
 import { once } from 'node:events';
 import { readdir, readFile } from 'node:fs/promises';
-import { dirname, join, relative } from 'node:path';
+import { basename, dirname, join, relative } from 'node:path';
 import { fileURLToPath } from 'node:url';
 import { Worker, isMainThread, workerData } from 'node:worker_threads';
 
-import { ActorSourceType } from '@apify/consts';
+import { ACTOR_SOURCE_TYPES } from '@apify/consts';
 import { cryptoRandomObjectId } from '@apify/utilities';
 import { ApifyClient } from 'apify-client';
 
-const basePath = join(dirname(fileURLToPath(import.meta.url)), 'sdk');
+const rootPath = dirname(fileURLToPath(import.meta.url));
+const basePath = join(rootPath, 'sdk');
 const actorBasePath = join(basePath, 'actorBase');
 
 async function run() {
     console.log(`Running E2E SDK tests`);
 
     const paths = await readdir(basePath, { withFileTypes: true });
-    const dirs = paths.filter((dirent) => dirent.isDirectory() && dirent.name !== actorBasePath.name);
+    const dirs = paths.filter((dirent) => dirent.isDirectory() && dirent.name !== basename(actorBasePath));
 
     for (const dir of dirs) {
         await runWorker(dir.name);
@@ -30,34 +31,59 @@ async function runWorker(dirName) {
         stderr: true,
     });
 
+    worker.on('exit', (exitCode) => {
+        if (exitCode !== 0) {
+            console.error(`Test ${dirName} failed`);
+            console.log('Captured stdout:');
+            worker.stdout.pipe(process.stdout);
+            console.log('Captured stderr:');
+            worker.stdout.pipe(process.stderr);
+        } else {
+            console.log(`${dirName} OK`);
+        }
+    });
+
     await once(worker, 'exit');
 }
 
-async function runTest(dirName) {
-    const testDir = join(basePath, dirName);
-
+async function packDir(dirName) {
     const sourceFiles = {};
     const textSuffixes = ['.mjs', '.json', 'Dockerfile'];
 
-    for (const dirent of await readdir(actorBasePath, { recursive: true, withFileTypes: true })) {
+    for (const dirent of await readdir(dirName, { recursive: true, withFileTypes: true })) {
         if (!dirent.isFile()) {
             continue;
         }
 
-        const name = relative(actorBasePath, join(dirent.parentPath, dirent.name));
-        const format = textSuffixes.some((suffix) => name.endsWith(suffix)) ? 'text' : 'base64';
+        const name = relative(dirName, join(dirent.parentPath, dirent.name));
+        const format = textSuffixes.some((suffix) => name.endsWith(suffix)) ? 'TEXT' : 'BASE64';
 
         sourceFiles[name] = {
             name,
             format,
             content: await readFile(
-                join(dirent.parentPath, dirent.name), 
-                { encoding: format === 'base64' ? 'base64' : 'utf8' }
+                join(dirent.parentPath, dirent.name),
+                { encoding: format === 'base64' ? 'base64' : 'utf8' },
             ),
         };
     }
 
-    // TODO obtain and include the built package
+    return sourceFiles;
+}
+
+async function runTest(dirName) {
+    const testDir = join(basePath, dirName);
+
+    const sourceFiles = {
+        ...(await packDir(actorBasePath)),
+        ...(await packDir(testDir)),
+    };
+
+    sourceFiles['apify.tgz'] = {
+        name: 'apify.tgz',
+        format: 'BASE64',
+        content: await readFile(join(rootPath, 'apify.tgz'), { encoding: 'base64' }),
+    };
 
     const client = new ApifyClient({
         token: process.env.APIFY_TOKEN,
@@ -67,25 +93,40 @@ async function runTest(dirName) {
 
     const actor = await client.actors().create({
         name: actorName,
-        defaultRunBuild: 'latest',
-        defaultRunMemoryMbytes: 256,
-        defaultRunTimeoutSecs: 600,
+        defaultRunOptions: {
+            build: 'latest',
+            memoryMbytes: 256,
+            timeoutSecs: 600,
+        },
         versions: [
             {
                 versionNumber: '0.0',
                 buildTag: 'latest',
-                sourceType: ActorSourceType.SOURCE_FILES,
-                sourceFiles,
+                sourceType: ACTOR_SOURCE_TYPES.SOURCE_FILES,
+                sourceFiles: Array.from(Object.values(sourceFiles)),
             },
         ],
     });
 
     const actorClient = client.actor(actor.id);
-    await actorClient.build({ versionNumber: '0.0' });
+    await actorClient.build('0.0', { waitForFinish: 30 * 60 * 1000 });
 
-    execFile(process.argv[0], [join(testDir, 'test.mjs'), actorName]);
+    const testProcessFinished = Promise.withResolvers();
+    const testProcess = spawn(
+        process.argv[0],
+        [join(testDir, 'test.mjs'), actor.id],
+        {stdio: 'inherit'},
+    );
 
-    await actorClient.delete();
+    testProcess.on('close', testProcessFinished.resolve);
+
+    await testProcessFinished.promise;
+
+    if (testProcess.exitCode === 0) {
+        await actorClient.delete();
+    }
+
+    process.exit(testProcess.exitCode);
 }
 
 if (isMainThread) {
